requirements.txt
python-telegram-bot[job-queue]==20.0
requests
matplotlib
seaborn
websockets
iqoptionapi
Pillow
